refactor: 모듈 분리 전 `JpaEntityTest` 상속 의존성 제거
