+++ src/test/java/kitchenpos/menu/domian/MenuTest.java
@@ -1,23 +1,39 @@
 package kitchenpos.menu.domian;
 
-import kitchenpos.JpaEntityTest;
-import kitchenpos.menu.domain.*;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuRepository;
-import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.product.domain.Product;
+import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.utils.DatabaseCleanup;
 import org.assertj.core.util.Lists;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.ActiveProfiles;
 
 import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+@DataJpaTest
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@ActiveProfiles("test")
+@Import(DatabaseCleanup.class)
 @DisplayName("메뉴 관련 도메인 테스트")
-public class MenuTest extends JpaEntityTest {
+public class MenuTest {
+    @Autowired
+    private DatabaseCleanup databaseCleanup;
+    @Autowired
+    private TestEntityManager entityManager;
 
     @Autowired
     private ProductRepository productRepository;
@@ -25,6 +41,12 @@ public class MenuTest extends JpaEntityTest {
     private MenuGroupRepository menuGroupRepository;
     @Autowired
     private MenuRepository menuRepository;
+
+
+    @BeforeEach
+    void setUp() {
+        databaseCleanup.execute();
+    }
 
     @DisplayName("메뉴 생성 테스트")
     @Test
@@ -64,5 +86,10 @@ public class MenuTest extends JpaEntityTest {
         // then
         assertThatThrownBy(() -> savedMenu.addMenuProducts(Lists.newArrayList(menuProduct)))
                 .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    private void flushAndClear() {
+        entityManager.flush();
+        entityManager.clear();
     }
 }
